refactor(empty): move to script setup
