+++ src/packages/__VUE/empty/doc.md
@@ -48,6 +48,14 @@ app.use(Empty);
 | image | 自定义图片 |
 | description | 自定义描述文字 |
 
+### 类型定义 version
+
+组件导出以下类型定义：
+
+```ts
+import type { EmptyImage, EmptyProps, EmptyInstance } from '@nutui/nutui';
+```
+
 ## 主题定制
 
 ### 样式变量
