+++ src/packages/__VUE/empty/index.ts
@@ -1,0 +1,12 @@
+import Empty from './empty.vue';
+import type { ComponentPublicInstance } from 'vue';
+import { withInstall } from '@/packages/utils';
+
+withInstall(Empty);
+
+export type { emptyProps } from './empty.vue';
+export type { EmptyImage } from './types';
+
+export type EmptyInstance = ComponentPublicInstance & InstanceType<typeof Empty>;
+
+export { Empty, Empty as default };
